Date: 5/14/18 Commit: Merged all of my work form my isolated project into the main project By: Michael J. Buck II

+++ ProjectKickAssSurvivalGame/Assets/Buck/Scripts/Turrets/BaseTurret.cs
@@ -11,7 +11,7 @@ public class BaseTurret : MonoBehaviour
 
     public float range;
 
-    public float damagePerShot;
+    public int damagePerShot;
 
     public float rotateSpeed;
